by hjq

+++ HomeworkWeb/src/main/java/com/fiverings/homeworkweb/jparepository/CourseJpaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fiverings.homeworkweb.model.Course;
 
-public interface CourseJpaRepository extends JpaRepository<Course, Long>{
+public interface CourseJpaRepository extends JpaRepository<Course, Integer>{
 	
 }
