+++ HomeworkWeb/src/main/java/com/fiverings/homeworkweb/jparepository/HomeworkJpaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fiverings.homeworkweb.model.Homework;
 
-public interface HomeworkJpaRepository extends JpaRepository<Homework, Long>{
+public interface HomeworkJpaRepository extends JpaRepository<Homework, Integer>{
 
 }
