+++ HomeworkWeb/src/main/java/com/fiverings/homeworkweb/jparepository/NoticeJpaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.fiverings.homeworkweb.model.Notice;
 
 
-public interface NoticeJpaRepository extends JpaRepository<Notice, Long>{
+public interface NoticeJpaRepository extends JpaRepository<Notice, Integer>{
 
 }
