+++ HomeworkWeb/src/main/java/com/fiverings/homeworkweb/jparepository/StudentJpaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fiverings.homeworkweb.model.Student;
 
-public interface StudentJpaRepository extends JpaRepository<Student,Long>{
+public interface StudentJpaRepository extends JpaRepository<Student,Integer>{
 
 }
