+++ HomeworkWeb/src/main/java/com/fiverings/homeworkweb/jparepository/TeacherJpaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fiverings.homeworkweb.model.Teacher;
 
-public interface TeacherJpaRepository extends JpaRepository<Teacher, Long>{
+public interface TeacherJpaRepository extends JpaRepository<Teacher, Integer>{
 
 }
